+++ db/migrate/20221227141529_add_reservable_to_machine.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# From this migration, the machines will be able to appear in the list, but without being reservable
+class AddReservableToMachine < ActiveRecord::Migration[5.2]
+  def change
+    add_column :machines, :reservable, :boolean, default: true
+  end
+end