+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog Fab-manager
 
 - Ability to group machines by categories
+- Ability to mark a machine as reservable or not
 - Private note on member's profile
 - Optional external identifier for users
 - Ability to disable generation of invoices at zero